+++ src/cli/Command/GenerateAction.js
