+++ src/cli/Command/GenerateMigration.js
