+++ src/cli/Command/GenerateModel.js
