+++ src/cli/Command/New.js
